# src/main/java/com/interjoin/teach/dtos/requests/ReviewRequest.java
package com.interjoin.teach.dtos.requests;

import lombok.Data;
import lombok.RequiredArgsConstructor;

@RequiredArgsConstructor
@Data
public class ReviewRequest {

    private String review;
    private String sessionUuid;
}