# src/main/java/com/interjoin/teach/dtos/responses/AuthResponse.java
package com.interjoin.teach.dtos.responses;

import com.interjoin.teach.dtos.UserDto;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder(toBuilder = true)
@NoArgsConstructor
@AllArgsConstructor
public class AuthResponse {

    private String token;
    private String refreshToken;
    private UserDto userDetails;

}