# src/main/resources/openapi.yaml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8082/api","description":"Generated server url"}],"paths":{"/book/list":{"put":{"tags":["session-controller"],"operationId":"getMyBookingSessions","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SessionDto"}}}}}}}},"/book/decline/{sessionUuid}":{"put":{"tags":["session-controller"],"operationId":"declineBookSession","parameters":[{"name":"sessionUuid","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/book/approve/{sessionUuid}":{"put":{"tags":["session-controller"],"operationId":"approveBookSession","parameters":[{"name":"sessionUuid","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/auth/reset":{"put":{"tags":["auth-controller"],"operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/auth/forgot":{"put":{"tags":["auth-controller"],"operationId":"forgotPassword","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/suggest":{"post":{"tags":["suggestion-controller"],"operationId":"save","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SuggestionDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/stripe/success-webhook":{"post":{"tags":["stripe-payment-controller"],"operationId":"handleSuccessfulPayment","parameters":[{"name":"Stripe-Signature","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/review":{"post":{"tags":["review-controller"],"operationId":"addReview","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/datat/available":{"post":{"tags":["test-controller"],"operationId":"test","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignupRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/datat/addsubjects":{"post":{"tags":["test-controller"],"operationId":"add","parameters":[{"name":"subjektet","in":"query","required":true,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK"}}}},"/book/new":{"post":{"tags":["session-controller"],"operationId":"bookNewSession","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookSessionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/auth/update-profile":{"post":{"tags":["auth-controller"],"operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/auth/signup/teacher":{"post":{"tags":["auth-controller"],"operationId":"signupTeacher","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SignupResponseDto"}}}}}}},"/auth/signup/student":{"post":{"tags":["auth-controller"],"operationId":"signupStudent","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SignupResponseDto"}}}}}}},"/auth/signup/agency":{"post":{"tags":["auth-controller"],"operationId":"signupStudent_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgencySignupRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/auth/signin":{"post":{"tags":["auth-controller"],"operationId":"signInUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSignInRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/auth/resend-otp/{cognitoUsername}":{"post":{"tags":["auth-controller"],"operationId":"resendOtp","parameters":[{"name":"cognitoUsername","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/auth/profile-pic":{"post":{"tags":["auth-controller"],"operationId":"setUserProfilePic","parameters":[{"name":"userUuid","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["pic"],"type":"object","properties":{"pic":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK"}}}},"/auth/cv":{"post":{"tags":["auth-controller"],"operationId":"uploadUserCv","parameters":[{"name":"userUuid","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["cv"],"type":"object","properties":{"cv":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK"}}}},"/auth/checkotp":{"post":{"tags":["auth-controller"],"operationId":"checkOtpCode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OtpVerifyRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/times/teacher/{teacherId}":{"get":{"tags":["times-controller"],"operationId":"getTeacherAvailableTimes","parameters":[{"name":"teacherId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AvailableTimesStringDto"}}}}}}}},"/times/specific-date/{teacherId}":{"get":{"tags":["times-controller"],"operationId":"getTeacherAvailableTimesForASpecificDate","parameters":[{"name":"teacherId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"date","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AvailableHourMinuteDto"}}}}}}}},"/times/signup":{"get":{"tags":["times-controller"],"operationId":"getAvailableTimes","parameters":[{"name":"timezone","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AvailableTimesSignupDto"}}}}}}},"/subcurr/all":{"get":{"tags":["subject-curriculum-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/SubjectCurriculumResponse"}}}}}}}},"/datat":{"get":{"tags":["test-controller"],"operationId":"test_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AvailableTimesStringDto"}}}}}}}},"/book/history":{"get":{"tags":["session-controller"],"operationId":"getCurrentStudentHistorySessions","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SessionDto"}}}}}}}},"/auth":{"get":{"tags":["auth-controller"],"operationId":"getUserDetails","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"delete":{"tags":["auth-controller"],"operationId":"deleteCurrentAccount","responses":{"200":{"description":"OK"}}}},"/auth/email":{"get":{"tags":["auth-controller"],"operationId":"checkIfUserEmailExists","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/agency/users":{"get":{"tags":["agency-controller"],"operationId":"getAgencyUsersPaginated","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageUserDto"}}}}}}},"/agency/teacher/{teacherId}":{"delete":{"tags":["agency-controller"],"operationId":"removeTeacherFromAgency","parameters":[{"name":"teacherId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"SessionDto":{"type":"object","properties":{"teacher":{"$ref":"#/components/schemas/UserDto"},"uuid":{"type":"string"},"studentFullName":{"type":"string"},"date":{"type":"string","format":"date-time"},"review":{"type":"string"},"reviewScore":{"type":"number","format":"double"}}},"SubjectCurriculumResponse":{"type":"object","properties":{"curriculum":{"type":"string"},"subjects":{"uniqueItems":true,"type":"array","items":{"type":"string"}}}},"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"location":{"type":"string"},"phoneNumber":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"profilePicture":{"type":"array","items":{"type":"string","format":"byte"}},"parentEmail":{"type":"string"},"shortBio":{"type":"string"},"longBio":{"type":"string"},"subCurrList":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/SubjectCurriculumResponse"}}}},"ResetPasswordDTO":{"required":["code","newPassword"],"type":"object","properties":{"email":{"type":"string"},"code":{"type":"string"},"newPassword":{"type":"string"}}},"SuggestionDto":{"type":"object","properties":{"type":{"type":"string"},"suggestion":{"type":"string"}}},"ReviewRequest":{"type":"object","properties":{"review":{"type":"string"},"sessionUuid":{"type":"string"}}},"AvailableTimesDto":{"type":"object","properties":{"weekDay":{"type":"string"},"availableTimes":{"type":"array","items":{"type":"string","format":"date-time"}}}},"SubjectCurriculumDto":{"type":"object","properties":{"curriculumName":{"type":"string"},"subjectNames":{"type":"array","items":{"type":"string"}}}},"UserSignupRequest":{"required":["dateOfBirth","email","firstName","lastName","password","timeZone"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"timeZone":{"type":"string"},"email":{"type":"string"},"location":{"type":"string"},"phoneNumber":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"parentEmail":{"type":"string"},"subCurrList":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/SubjectCurriculumDto"}},"shortBio":{"type":"string"},"longBio":{"type":"string"},"password":{"type":"string"},"availableTimes":{"type":"array","items":{"$ref":"#/components/schemas/AvailableTimesDto"}},"experiences":{"type":"array","items":{"type":"string"}},"qualifications":{"type":"string"},"experience":{"type":"string"},"agencyReferalCode":{"type":"string"}}},"AvailableHourMinuteDto":{"type":"object","properties":{"hourMinuteString":{"type":"string"},"dateTime":{"type":"string","format":"date-time"},"dateOfSession":{"type":"string","format":"date"}}},"BookSessionRequest":{"required":["date","teacherId"],"type":"object","properties":{"date":{"$ref":"#/components/schemas/AvailableHourMinuteDto"},"teacherId":{"type":"integer","format":"int64"},"subject":{"type":"string"},"curriculum":{"type":"string"}}},"UpdateProfileRequest":{"type":"object","properties":{"password":{"type":"string"},"location":{"type":"string"},"phone":{"type":"string"},"subCurrList":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/SubjectCurriculumDto"}}}},"SignupResponseDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"cognitoUsername":{"type":"string"},"uuid":{"type":"string"},"subCurrList":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/SubjectCurriculumDto"}}}},"AgencySignupRequest":{"required":["agencyName","contactEmail"],"type":"object","properties":{"agencyName":{"type":"string"},"contactEmail":{"type":"string"},"location":{"type":"string"},"numberOfTeachers":{"type":"integer","format":"int32"},"additionalComments":{"type":"string"}}},"UserSignInRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"token":{"type":"string"},"refreshToken":{"type":"string"},"userDetails":{"$ref":"#/components/schemas/UserDto"}}},"OtpVerifyRequest":{"type":"object","properties":{"cognitoUsername":{"type":"string"},"otpCode":{"type":"string"}}},"AvailableTimesStringDto":{"type":"object","properties":{"weekDay":{"type":"string"},"availableHourMinute":{"type":"array","items":{"$ref":"#/components/schemas/AvailableHourMinuteDto"}}}},"AvailableTimesSignupDto":{"type":"object","properties":{"availableTimes":{"type":"array","items":{"$ref":"#/components/schemas/AvailableHourMinuteDto"}}}},"Pageable":{"type":"object","properties":{"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"},"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/Sort"}}},"Sort":{"type":"object","properties":{"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageUserDto":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"pageable":{"$ref":"#/components/schemas/Pageable"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}},"numberOfElements":{"type":"integer","format":"int32"},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/Sort"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}}}}}